[Windows] Touchscreen window resize bug bandaid
Not sure if this was caused by a windows update or if it's just gone unnoticed til now...
After resizing the Touchscreen window, any Windows10(/8?) app will stop drawing passed the initial window size when resized.
I was only able to fix the issue by logging out/in or power cycling.
I'm guessing this is caused by calling resize() inside TouchScreen's resize event but this seems to be the only way to maintain aspect ratio...